pyproject.toml
[project]
name = "executive-assistant-negotiation-env"
version = "2.0.0"
description = "An OpenEnv environment for multi-agent executive assistant planning"
requires-python = ">=3.11"
dependencies = [
    "pydantic",
    "fastapi",
    "uvicorn",
    "openenv-core>=0.2.3",
    "pillow",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"
